# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lumen — AI-powered wellness companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sleep:
    get:
      operationId: listSleep
      tags: [sleep]
      summary: List sleep entries
      responses:
        "200":
          description: Sleep entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SleepEntry"
    post:
      operationId: createSleep
      tags: [sleep]
      summary: Log sleep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSleepBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepEntry"
  /sleep/{id}:
    delete:
      operationId: deleteSleep
      tags: [sleep]
      summary: Delete sleep entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    SleepEntry:
      type: object
      properties:
        id: { type: string }
        date: { type: string }
        durationHours: { type: number }
        quality:
          type: string
          enum: [poor, fair, good, excellent]
        bedtime: { type: string }
        wakeTime: { type: string }
        deepSleepHours:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required: [id, date, durationHours, quality, bedtime, wakeTime]
    CreateSleepBody:
      type: object
      properties:
        date: { type: string }
        durationHours: { type: number }
        quality:
          type: string
          enum: [poor, fair, good, excellent]
        bedtime: { type: string }
        wakeTime: { type: string }
        deepSleepHours:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required: [date, durationHours, quality, bedtime, wakeTime]
...